Split the wasm pipeline into linker and native-rung units
